build: fix Linux CI failures in tapyrus-genesis link and Qt6 notificator
CalculateSequenceLocks()/EvaluateSequenceLocks()/SequenceLocks() move out of
consensus/tx_verify.cpp into a new consensus/sequence_locks.cpp: they're the
only functions in that file touching CBlockIndex (GetAncestor(), defined in
chain.cpp, part of tapyrus_chainstate). tapyrus-genesis compiles its own copy
of consensus/tx_verify.cpp directly instead of linking tapyrus_chainstate;
Linux's linker requires every symbol referenced anywhere in a directly-linked
object file to resolve, even in functions nothing calls, so the unresolved
CBlockIndex::GetAncestor() reference broke the Linux build (macOS's linker
tolerated it). tapyrus_chainstate gains the new file alongside tx_verify.cpp,
unchanged for every other consumer.

notificator.cpp: FreedesktopImage::metaType() now returns QMetaType instead
of int, matching qDBusRegisterMetaType<T>()'s Qt6 return type (it returned
int in Qt5) and QVariant's Qt6 QMetaType-based constructor. Unrelated to the
above, just newly reachable now that the Linux build gets further than
before.
