+++ src/qt/notificator.cpp
@@ -78,7 +78,7 @@ public:
     FreedesktopImage() {}
     explicit FreedesktopImage(const QImage &img);
 
-    static int metaType();
+    static QMetaType metaType();
 
     // Image to variant that can be marshalled over DBus
     static QVariant toVariant(const QImage &img);
@@ -141,7 +141,7 @@ const QDBusArgument &operator>>(const QDBusArgument &a, FreedesktopImage &i)
     return a;
 }
 
-int FreedesktopImage::metaType()
+QMetaType FreedesktopImage::metaType()
 {
     return qDBusRegisterMetaType<FreedesktopImage>();
 }
